requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
python-telegram-bot==20.7
asyncio==3.4.3
logging==0.4.9.6
typing-extensions==4.8.0
json5==0.9.14
datetime==5.3